Move laske serialization code to its own module
